Redraw from scratch every 200 updates
This is required in order for mprdb to play nicely with dumb spectation
systems, such as termcast and dgamelaunch.  200 is an empirical number
chosen such as to not quite fill up termcast's 100KB output buffer;
dgamelaunch has an unlimited output buffer, but takes linear space,
so anything much larger than 200 would be unwise.